requirements.txt
langgraph>=0.2.0
python-dotenv>=1.0.0
langchain-google-genai>=2.0.0
